Changed how axe upgrades work. Now instead of a static value of 97 logs before going to the exchange--we grab the current price of logs from OSBuddy Exchange and calculate how many more logs we need before we can purchase axes. Progressive mode will not switch over to oak trees until we have enough logs to purchase axe upgrades.
If for whatever reason the OSBuddy Exchange goes offline or is undermaintenance, we will use a static value of 31gp/log to calculate how many logs we need to purchase axes. This value is tried and true over a 6 month period and should remain stable if logs should crash from their highly inflated price of 150gp+/log right now

Also in the Main class I've changed the screenshot class to only take daily screenshots of connection lost. I think this will be a toggle feature in the next update since it's not very useful for bot armies.